perf(desktop): lazy-load workspace focus reconciliation
Problem: merging the latest main-v2 provider bridge additions pushed the desktop initial raw bundle above its enforced 2200 KiB budget.\n\nRoot cause: the workspace focus and visibility reconciliation lifecycle was statically bundled even though the workspace refresh store was already loaded on demand.\n\nFix: move the lifecycle coordinator into the lazy workspace store, retain stale-load and cleanup fencing in App, and keep the project-tree subscription in the same refresh lifecycle.\n\nVerification: pnpm test:all, pnpm build, targeted Go race tests, repolint, and git diff --check.
